feat: ID'd connections for typing state

+++ src/app/api/[[...slugs]]/ws/storage/typing-state.ts
@@ -1,11 +1,20 @@
-export const typingStateStorage = new Set<string>();
+export const typingStateStorage = new Map<string, Set<string>>();
 
-export function tryAddTypingState(userId: string) {
-  typingStateStorage.add(userId);
+export function tryAddTypingState(userId: string, connectionId: string) {
+  if (!typingStateStorage.has(userId)) {
+    typingStateStorage.set(userId, new Set());
+  }
+  typingStateStorage.get(userId)?.add(connectionId);
 }
 
-export function tryRemoveTypingState(userId: string) {
-  typingStateStorage.delete(userId);
+export function tryRemoveTypingState(userId: string, connectionId: string) {
+  const userConnections = typingStateStorage.get(userId);
+  if (userConnections) {
+    userConnections.delete(connectionId);
+    if (userConnections.size === 0) {
+      typingStateStorage.delete(userId);
+    }
+  }
 }
 
 export function getTypingState(userId: string) {
